Build with scikit-build-core
