+++ pyproject.toml
@@ -1,0 +1,50 @@
+[build-system]
+requires = [
+    "scikit-build-core",
+]
+
+build-backend = "scikit_build_core.build"
+
+[project]
+name = "walnuts"
+description = "The Within-Orbit Adaptive No-U-Turn Sampler"
+readme = "README.md"
+license = { text = "BSD-3-Clause" }
+authors = [
+    { name = "Brian Ward", email = "bward@flatironinstitute.org" },
+    { name = "Bob Carpenter", email = "bcarpenter@flatironinstitute.org" },
+]
+requires-python = ">=3.9"
+dependencies = ["numpy"]
+dynamic = ["version"]
+
+
+[tool.scikit-build]
+minimum-version = "0.10"
+cmake.source-dir = "."
+build.targets = ["walnutpie"]
+cmake.args = ["-G Ninja"]
+wheel.packages=["python/src/walnuts"]
+# Indicate that we don't depend on the CPython API
+wheel.py-api = "py3"
+build-dir = "build/{wheel_tag}"
+editable.rebuild=true
+
+build.verbose = true
+logging.level = "INFO"
+
+[tool.scikit-build.metadata.version]
+# Instead of hardcoding the version here, extract it from the source files.
+provider = "scikit_build_core.metadata.regex"
+input = "python/src/walnuts/__init__.py"
+
+[project.urls]
+"Homepage" = "https://github.com/flatironinstitute/walnuts"
+"Bug Tracker" = "https://github.com/flatironinstitute/walnuts/issues"
+
+[project.optional-dependencies]
+stan = ["bridgestan>=2.9.0", "stanio"]
+compiled = ["numba"]
+
+[tool.isort]
+profile = "black"
